+++ CHANGELOG.md
@@ -9,9 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - include a setuptools `setup.py` script for potential distribution
 - git ignore for python projects
 - documentation to `sample_config.ini` file
+- support rebuild of packages (must be specifed on command line)
 
 ## Changed
 - use subparser of `argparse` - makes subcommand handling cleaner
+- improve list print out
+- improve error message on command
 
 ## [1.0.0] - 2017-06-20
 ### Added